+++ CHANGELOG.md
@@ -12,8 +12,10 @@
 	- Currently no non-footwear items are in that tag, but now you can add them yourself
 - The Biome Painter now defaults to having a radius of 2 instead of 0
   - Now it has a default volume of 5x5x5 instead of 1x1x1
+- Ender Anchors now have a Block Entity, which allows them to do some more stuff
 - When holding a Portable Ender Bridge, Ender Anchors now glow through blocks
 - The Portable Ender Bridge now lets you teleport to Ender Anchors even if there's a block in the way
+- The Portable Ender Bridge makes a sound when used
 
 ### Fixed